complete merge, add a test
